feat(accordéons): ajout d'ancre sur tous les accordéons pour pouvoir les ouvrir en mettant une ancre dans l'url
