+++ package.json
@@ -9,10 +9,10 @@
     "build:npm": "lerna run build --scope=@socialgouv/cdtn-utils --scope=@socialgouv/cdtn-ui --scope=@socialgouv/modeles-social --stream",
     "clean": "lerna clean",
     "dev:docker": "docker-compose up elasticsearch",
-    "dev": "lerna run dev --parallel",
+    "dev": "SENTRY_IGNORE_API_RESOLUTION_ERROR=1 lerna run dev --parallel",
     "type-check": "lerna run type-check --stream",
     "dev:frontend": "lerna run dev --parallel --scope=@cdt/frontend --scope=@socialgouv/cdtn-ui --scope=@socialgouv/modeles-social --stream",
-    "dev:frontend:preprod-api": "NEXT_PUBLIC_APP_ENV=external-api NEXT_PUBLIC_SITE_URL=https://code-du-travail-numerique-preprod.ovh.fabrique.social.gouv.fr yarn dev:frontend",
+    "dev:frontend:preprod-api": "SENTRY_IGNORE_API_RESOLUTION_ERROR=1  NEXT_PUBLIC_APP_ENV=external-api NEXT_PUBLIC_SITE_URL=https://code-du-travail-numerique-preprod.ovh.fabrique.social.gouv.fr yarn dev:frontend",
     "dev:simulate-deployment": "NEXT_PUBLIC_APP_ENV=external-api NEXT_PUBLIC_SITE_URL=https://code-du-travail-numerique-preprod.ovh.fabrique.social.gouv.fr yarn workspace @cdt/frontend build && yarn workspace @cdt/frontend start",
     "lint": "lerna run lint",
     "lint:fix": "lerna run lint -- --fix",
