+++ packages/code-du-travail-frontend/src/contributions/DisplayContentContribution.tsx
@@ -1,10 +1,5 @@
-import {
-  Accordion,
-  Alert,
-  Heading,
-  Table as UITable,
-  theme,
-} from "@socialgouv/cdtn-ui";
+import { Alert, Heading, Table as UITable, theme } from "@socialgouv/cdtn-ui";
+import { AccordionWithAnchor as Accordion } from "../../src/common/AccordionWithAnchor";
 
 import styled from "styled-components";
 import { FicheServicePublic } from "../fiche-service-public";
